Fix type of array.

+++ python/boomer/common/cpp/losses.cpp
@@ -4,7 +4,7 @@
 using namespace losses;
 
 
-DefaultPrediction::DefaultPrediction(intp numPredictions, intp* predictedScores) {
+DefaultPrediction::DefaultPrediction(intp numPredictions, float64* predictedScores) {
     numPredictions_ = numPredictions;
     predictedScores_ = predictedScores;
 }
@@ -13,12 +13,12 @@ DefaultPrediction::~DefaultPrediction() {
     free(predictedScores_);
 }
 
-Prediction::Prediction(intp numPredictions, intp* predictedScores, float64 overallQualityScore)
+Prediction::Prediction(intp numPredictions, float64* predictedScores, float64 overallQualityScore)
     : DefaultPrediction(numPredictions, predictedScores) {
     overallQualityScore_ = overallQualityScore;
 }
 
-LabelWisePrediction::LabelWisePrediction(intp numPredictions, intp* predictedScores, float64* qualityScores,
+LabelWisePrediction::LabelWisePrediction(intp numPredictions, float64* predictedScores, float64* qualityScores,
                                          float64 overallQualityScore)
     : Prediction(numPredictions, predictedScores, overallQualityScore) {
     qualityScores_ = qualityScores;
